requirements.txt
dnspython>=2.1.0
python-nmap>=0.7.1
requests>=2.26.0
shodan>=1.27.0
beautifulsoup4>=4.10.0
pyyaml>=6.0
tqdm>=4.62.3
cryptography>=36.0.0